chore: checkstyle, ref: package rename
