+++ src/main/java/com/troblecodings/invisiblelights/blocks/BlockCustomLight.java
@@ -1,0 +1,8 @@
+package com.troblecodings.invisiblelights.blocks;
+
+public class BlockCustomLight extends BlockInvisibleLight {
+
+    public BlockCustomLight(final int light) {
+        super(light);
+    }
+}
