+++ src/main/java/com/troblecodings/invisiblelights/blocks/BlockLightBlocker.java
@@ -1,0 +1,18 @@
+package com.troblecodings.invisiblelights.blocks;
+
+import net.minecraft.core.BlockPos;
+import net.minecraft.world.level.BlockGetter;
+import net.minecraft.world.level.block.state.BlockState;
+
+public class BlockLightBlocker extends BlockInvisibleLight {
+
+    public BlockLightBlocker() {
+        super(0);
+    }
+
+    @Override
+    public int getLightBlock(final BlockState state, final BlockGetter worldIn,
+            final BlockPos pos) {
+        return 15;
+    }
+}
